+++ src/test/java/com/infoshare/academy/MainDBTest.java
@@ -1,7 +1,7 @@
 package com.infoshare.academy;
 
 import com.infoshare.academy.database.CarsDAO;
-import com.infoshare.academy.model.cars.Car;
+import com.infoshare.academy.model.cars.entity.Car;
 import com.infoshare.academy.model.carsservice.CarDescription;
 
 import java.util.List;
